+++ Models/Pomodoro/Timer/PomodoroIntervalModel.cs
@@ -6,6 +6,6 @@ public record PomodoroIntervalModel
 {
     public required ECycleType Type { get; init; } = ECycleType.Pomodoro;
     public bool Done { get; set; } = false;
-    public required int TimeInMinutes { get; init; } = 25;
+    public required double TimeInMinutes { get; init; } = 25;
     public required int Order { get; set; } = 1;
 }
